backend/requirements.txt
altair==5.4.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
narwhals==1.14.1
referencing==0.35.1
rpds-py==0.21.0
Flask-SQLAlchemy==3.1.1
Jinja2==3.1.4
MarkupSafe==3.0.2
Werkzeug==3.1.3
blinker==1.9.0
click==8.1.7
flask==3.1.0
importlib-metadata==8.5.0
itsdangerous==2.2.0
sqlalchemy==2.0.36
typing-extensions==4.12.2
zipp==3.21.0
Flask-JWT-Extended==4.7.1
PyJWT==2.10.0
Flask-Migrate==4.0.7
Mako==1.3.6
alembic==1.14.0
Flask-Cors==5.0.0
annotated-types==0.7.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api==python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
httplib2==0.22.0
idna==3.10
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.0
pydantic-core==2.27.0
pyparsing==3.2.0
requests==2.32.3
rsa==4.9
tqdm==4.67.0
uritemplate==4.1.1
urllib3==2.2.3
GitPython==3.1.43
gitdb==4.0.11
smmap==5.0.1
chardet==5.2.0
PyPDF2==3.0.1
PyYAML==6.0.2
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.7
langchain-core==0.3.19
langchain-text-splitters==0.3.2
langsmith==0.1.144
multidict==6.1.0
numpy==1.26.4
orjson==3.10.11
packaging==24.2
propcache==0.2.0
requests-toolbelt==1.0.0
sniffio==1.3.1
tenacity==9.0.0
yarl==1.18.0
langchain-google-genai==2.0.5
SQLAlchemy==2.0.35
dataclasses-json==0.6.7
httpx-sse==0.4.0
langchain-community==0.3.7
marshmallow==3.23.1
mypy-extensions==1.0.0
pydantic-settings==2.6.1
python-dotenv==1.0.1
typing-inspect==0.9.0
faiss-cpu==1.9.0.post1
celery~=5.4.0
python-Levenshtein
scikit-learn
